requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bower.py==1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-extensions==3.2.1
django-jquery==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
executing==1.2.0
github3.py==3.2.0
idna==3.4
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
semantic-version==2.10.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
